feat: allow connection to mcp server via a jwt
