+++ api/api/services/security_service_test.py
@@ -1,3 +1,6 @@
+# pyright: reportPrivateUsage=false
+
+import os
 from unittest.mock import Mock, patch
 
 import pytest
@@ -5,7 +8,7 @@ from fastapi import HTTPException
 
 from api.dependencies.security import UserClaims
 from api.services.analytics._analytics_service import AnalyticsService
-from api.services.security_service import SecurityService
+from api.services.security_service import SecurityService, _default_key_ring
 from core.domain.errors import InvalidToken
 from core.domain.events import EventRouter
 from core.domain.tenant_data import TenantData
@@ -387,3 +390,31 @@ class TestUserOrganization:
         )
         mock_org_storage.create_organization.assert_called_once_with(expected_org)
         assert result == expected_org
+
+
+class TestDefaultKeyRing:
+    @pytest.fixture(scope="function", autouse=True)
+    def patch_jwks_url(self):
+        with patch.dict(os.environ, {"WORKFLOWAI_JWKS_URL": "https://hello"}, clear=True):
+            yield
+
+    def test_not_present(self):
+        assert "WORKFLOWAI_JWK" not in os.environ, "sanity"
+
+        kr = _default_key_ring()
+        assert kr.key_cache == {}
+
+    def test_present_not_valid(self):
+        assert "WORKFLOWAI_JWK" not in os.environ, "sanity"
+        # Sending invalid value in WORKFLOWAI_JWK
+        os.environ["WORKFLOWAI_JWK"] = "hello"
+
+        kr = _default_key_ring()
+        assert kr.key_cache == {}
+
+    def test_present_valid(self):
+        os.environ["WORKFLOWAI_JWK"] = (
+            "eyJrdHkiOiJFQyIsIngiOiJLVUpZYzd2V0R4Um55NW5BdC1VNGI4MHRoQ1ZuaERUTDBzUmZBRjR2cDdVIiwieSI6IjM0dWx1VDgyT0RFRFJXVU9KNExrZzFpanljclhqMWc1MmZRblpqeFc5cTAiLCJjcnYiOiJQLTI1NiIsImlkIjoiMSJ9Cg=="
+        )
+        kr = _default_key_ring()
+        assert len(kr.key_cache) == 1
